+++ src/gamspy/_model.py
@@ -1149,7 +1149,9 @@ class Model:
 
         """
         if solver is None:
-            solver = utils.getDefaultSolvers()[str(self.problem).upper()]
+            solver = utils.getDefaultSolvers(self.container.system_directory)[
+                str(self.problem).upper()
+            ]
 
         validation.validate_solver_args(
             self.container.system_directory,