+++ src/gamspy/utils.py
@@ -52,15 +52,45 @@ elif platform.system() == "Darwin":
 elif platform.system() == "Windows":
     DEFAULT_DIR = os.path.join(user_dir, "Documents", "GAMSPy")
 
+_defaults: dict[str, str] | None = None
+_capabilities: dict[str, list[str]] | None = None
+_installed_solvers: list[str] | None = None
 
-def getDefaultSolvers() -> dict[str, str]:
-    try:
-        import gamspy_base
-    except ModuleNotFoundError as e:  # pragma: no cover
-        e.msg = "You must first install gamspy_base to use this functionality"
-        raise e
 
-    return gamspy_base.defaults
+def getDefaultSolvers(system_directory: str) -> dict[str, str]:
+    """
+    Returns the default solver for each problem type.
+
+    Parameters
+    ----------
+    system_directory : str
+
+    Returns
+    -------
+    dict[str, str]
+
+    Examples
+    --------
+    >>> import gamspy as gp
+    >>> import gamspy_base
+    >>> default_solvers = gp.utils.getDefaultSolvers(gamspy_base.directory)
+
+    """
+    global _defaults
+    if _defaults is not None:
+        return _defaults
+
+    capabilities_path = os.path.join(system_directory, CAPABILITIES_FILE)
+    with open(capabilities_path, encoding="utf-8") as file:
+        lines = file.read().split("DEFAULTS")[1].splitlines()[1:]
+
+    defaults: dict[str, str] = dict()
+    for line in lines:
+        problem, solver = line.split()
+        defaults[problem] = solver
+
+    _defaults = defaults
+    return _defaults
 
 
 def getSolverCapabilities(system_directory: str) -> dict[str, list[str]]:
@@ -76,6 +106,10 @@ def getSolverCapabilities(system_directory: str) -> dict[str, list[str]]:
     -------
     dict[str, list[str]]
     """
+    global _capabilities
+    if _capabilities is not None:
+        return _capabilities
+
     capabilities_path = os.path.join(system_directory, CAPABILITIES_FILE)
     capabilities: dict[str, list[str]] = dict()
 
@@ -96,7 +130,8 @@ def getSolverCapabilities(system_directory: str) -> dict[str, list[str]]:
 
         capabilities[solver] = problem_types
 
-    return capabilities
+    _capabilities = capabilities
+    return _capabilities
 
 
 def getInstalledSolvers(system_directory: str) -> list[str]:
@@ -119,6 +154,10 @@ def getInstalledSolvers(system_directory: str) -> list[str]:
     >>> installed_solvers = utils.getInstalledSolvers(gamspy_base.directory)
 
     """
+    global _installed_solvers
+    if _installed_solvers is not None:
+        return _installed_solvers
+
     capabilities_path = os.path.join(system_directory, CAPABILITIES_FILE)
     solvers: list[str] = []
 
@@ -140,7 +179,8 @@ def getInstalledSolvers(system_directory: str) -> list[str]:
         if solver != "GUSS":
             solvers.append(solver)
 
-    return sorted(solvers)
+    _installed_solvers = sorted(solvers)
+    return _installed_solvers
 
 
 def getAvailableSolvers() -> list[str]: