+++ tests/unit/test_utils.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+import time
+
 import pytest
 
 import gamspy.utils as utils
@@ -87,7 +89,9 @@ def test_available_solvers(data):
 
 
 def test_default_solvers():
-    default_solvers = utils.getDefaultSolvers()
+    import gamspy_base
+
+    default_solvers = utils.getDefaultSolvers(gamspy_base.directory)
 
     expected = {
         "CNS": "PATH",
@@ -107,3 +111,37 @@ def test_default_solvers():
     }
 
     assert default_solvers == expected
+
+
+def test_solver_and_capability_caching():
+    import gamspy_base
+
+    start = time.perf_counter_ns()
+    _ = utils.getDefaultSolvers(gamspy_base.directory)
+    first_time = time.perf_counter_ns() - start
+
+    start = time.perf_counter_ns()
+    _ = utils.getDefaultSolvers(gamspy_base.directory)
+    second_time = time.perf_counter_ns() - start
+
+    assert second_time < first_time, f"{first_time=}, {second_time=}"
+
+    start = time.perf_counter_ns()
+    _ = utils.getSolverCapabilities(gamspy_base.directory)
+    first_time = time.perf_counter_ns() - start
+
+    start = time.perf_counter_ns()
+    _ = utils.getSolverCapabilities(gamspy_base.directory)
+    second_time = time.perf_counter_ns() - start
+
+    assert second_time < first_time, f"{first_time=}, {second_time=}"
+
+    start = time.perf_counter_ns()
+    _ = utils.getInstalledSolvers(gamspy_base.directory)
+    first_time = time.perf_counter_ns() - start
+
+    start = time.perf_counter_ns()
+    _ = utils.getInstalledSolvers(gamspy_base.directory)
+    second_time = time.perf_counter_ns() - start
+
+    assert second_time < first_time, f"{first_time=}, {second_time=}"